Revert "Feature/json ld try"
