+++ .env.dist
@@ -11,6 +11,6 @@ APP_SECRET=0a4f68d47f8d7020d43986c93dd7d3e2
 ###> doctrine/doctrine-bundle ###
 # Format described at http://docs.doctrine-project.org/projects/doctrine-dbal/en/latest/reference/configuration.html#connecting-using-a-url
 # For an SQLite database, use: "sqlite:///%kernel.project_dir%/var/data.db"
-# Configure your db driver and server_version in config/packages/doctrine.yaml
-DATABASE_URL=sqlite:///%kernel.project_dir%/var/data/main.db
+# Set "serverVersion" to your server version to avoid edge-case exceptions and extra database calls
+DATABASE_URL="mysql://root@127.0.0.1:3306/symfony?charset=utf8mb4&serverVersion=5.7"
 ###< doctrine/doctrine-bundle ###
