--- composer-flex.json
@@ -1,63 +1,0 @@
-{
-    "type": "project",
-    "license": "proprietary",
-    "require": {
-        "php": "^7.0.8",
-        "ext-pdo_sqlite": "*",
-        "ext-curl": "*",
-        "api-platform/api-pack": "^1.0",
-        "doctrine/doctrine-migrations-bundle": "^1.3",
-        "doctrine/orm": "^2.4.5",
-        "php-http/httplug-pack": "^1.0",
-        "sensio/framework-extra-bundle": "^5.1.0",
-        "symfony/console": "^4.0.0",
-        "symfony/framework-bundle": "^4.0.0",
-        "symfony/monolog-bundle": "^3.1.0",
-        "symfony/profiler-pack": "^1.0",
-        "symfony/yaml": "^4.0.0"
-    },
-    "require-dev": {
-        "symfony/webpack-encore-pack": "^1.0",
-        "symfony/flex": "^1.0",
-        "symfony/dotenv": "^4.0.0"
-    },
-    "config": {
-        "preferred-install": {
-            "*": "dist"
-        },
-        "sort-packages": true
-    },
-    "autoload": {
-        "psr-4": {
-            "App\\": "src/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "App\\Tests\\": "tests/"
-        }
-    },
-    "scripts": {
-        "auto-scripts": {
-            "cache:clear": "symfony-cmd",
-            "assets:install --symlink --relative %PUBLIC_DIR%": "symfony-cmd"
-        },
-        "post-install-cmd": [
-            "@auto-scripts"
-        ],
-        "post-update-cmd": [
-            "@auto-scripts"
-        ]
-    },
-    "conflict": {
-        "symfony/symfony": "*",
-        "symfony/twig-bundle": "<3.3",
-        "symfony/debug": "<3.3"
-    },
-    "extra": {
-        "symfony": {
-            "id": "01BXPDHBAP11MNR97CR7E97D3F",
-            "allow-contrib": false
-        }
-    }
-}
