+++ composer.json
@@ -3,12 +3,8 @@
     "license": "proprietary",
     "require": {
         "php": "^7.0.8",
-        "ext-curl": "*",
         "ext-pdo_sqlite": "*",
-        "doctrine/doctrine-migrations-bundle": "^1.3",
-        "javiereguiluz/easyadmin-bundle": "^1.17",
-        "php-http/httplug-pack": "^1.0",
-        "rebolon/api-pack": "^1.0.2",
+        "api-platform/api-pack": "^1.0",
         "sensio/framework-extra-bundle": "^5.1.0",
         "symfony/console": "^4.0.0",
         "symfony/framework-bundle": "^4.0.0",
@@ -17,18 +13,10 @@
         "symfony/yaml": "^4.0.0"
     },
     "require-dev": {
-        "symfony/dotenv": "^4.0.0",
+        "symfony/webpack-encore-pack": "^1.0",
         "symfony/flex": "^1.0",
-        "symfony/webpack-encore-pack": "^1.0"
+        "symfony/dotenv": "^4.0.0"
     },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "repositories": [
-        {
-            "type": "vcs",
-            "url":  "git@github.com:Rebolon/api-pack.git"
-        }
-    ],
     "config": {
         "preferred-install": {
             "*": "dist"
