--- config/packages/httplug.yaml
@@ -1,14 +1,0 @@
-httplug:
-    plugins:
-        redirect:
-            preserve_header: true
-
-    discovery:
-        client: 'auto'
-
-    clients:
-        app:
-            http_methods_client: true
-            plugins:
-                - 'httplug.plugin.content_length'
-                - 'httplug.plugin.redirect'
