--- config/packages/twig_extensions.yaml
@@ -1,10 +1,0 @@
-services:
-    _defaults:
-        public: false
-        autowire: true
-        autoconfigure: true
-
-    #Twig\Extensions\ArrayExtension: ~
-    #Twig\Extensions\DateExtension: ~
-    #Twig\Extensions\IntlExtension: ~
-    #Twig\Extensions\TextExtension: ~
