+++ package.json
@@ -12,11 +12,11 @@
     },
     "scripts": {
         "init-project": "npm run init-project:linux",
-        "init-project:linux": "cp .env.dist .env && composer install && npm install && npm run db-init:linux",
-        "init-project:windows": "cp .env.dist .env && composer install && npm install && npm run db-init:windows",
+        "init-project:linux": "composer install && npm install &&  && npm run db-init:linux",
+        "init-project:windows": "composer install && npm install && npm run db-init:windows",
         "db-init": "npm run db-init:linux",
-        "db-init:linux": "$npm_package_config_php bin/console doctrine:database:create && $npm_package_config_php bin/console doctrine:schema:create && $npm_package_config_php bin/console doctrine:migrations:migrate",
-        "db-init:windows": "%npm_package_config_php% bin/console doctrine:database:create && %npm_package_config_php% bin/console doctrine:schema:create && %npm_package_config_php% bin/console doctrine:migrations:migrate",
+        "db-init:linux": "$npm_package_config_php bin/console doctrine:database:create && $npm_package_config_php bin/console doctrine:schema:create",
+        "db-init:windows": "%npm_package_config_php% bin/console doctrine:database:create && %npm_package_config_php% bin/console doctrine:schema:create",
         "sf-dev": "npm run sf-dev:linux",
         "sf-dev:linux": "$npm_package_config_php -S 127.0.0.1:$npm_package_config_server_port_web -t public",
         "sf-dev:windows": "%npm_package_config_php% -S 127.0.0.1:%npm_package_config_server_port_web% -t public",
