--- src/Entity/Library/ProjectBookCreation.php
@@ -1,101 +1,0 @@
-<?php
-namespace App\Entity\Library;
-
-use ApiPlatform\Core\Annotation\ApiProperty;
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity
- * @ORM\Table(name="project_book_creation")
- */
-class ProjectBookCreation
-{
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="integer", nullable=false)
-     */
-    private $role;
-
-    /**
-     * @ORM\ManyToOne(
-     *     targetEntity="App\Entity\Library\Book",
-     *     inversedBy="projectBookCreation",
-     *     fetch="EAGER"
-     * )
-     * @ORM\JoinColumn(name="book_id", referencedColumnName="id")
-     */
-    private $book;
-
-    /**
-     * @ORM\ManyToOne(
-     *     targetEntity="App\Entity\Library\Author",
-     *     inversedBy="projectBookCreation",
-     *     fetch="EAGER"
-     * )
-     * @ORM\JoinColumn(name="author_id", referencedColumnName="id")
-     */
-    private $author;
-
-    /**
-     * @return mixed
-     */
-    public function getRole()
-    {
-        return $this->role;
-    }
-
-    /**
-     * @param mixed $role
-     * @return ProjectBookCreation
-     */
-    public function setRole($role)
-    {
-        $this->role = $role;
-
-        return $this;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getBook()
-    {
-        return $this->book;
-    }
-
-    /**
-     * @param Book $book
-     * @return $this
-     */
-    public function setBook(Book $book)
-    {
-        $this->book = $book;
-
-        return $this;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getAuthor()
-    {
-        return $this->author;
-    }
-
-    /**
-     * @param Author $author
-     * @return $this
-     */
-    public function setAuthor(Author $author)
-    {
-        $this->author = $author;
-
-        return $this;
-    }
-}
