--- src/Entity/Library/Serie.php
@@ -1,82 +1,0 @@
-<?php
-namespace App\Entity\Library;
-
-use ApiPlatform\Core\Annotation\ApiProperty;
-use ApiPlatform\Core\Annotation\ApiResource;
-use ApiPlatform\Core\Annotation\ApiSubresource;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ApiResource(iri="http://schema.org/Series")
- * @ORM\Entity
- */
-class Serie
-{
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    private $id;
-
-    /**
-     * @ApiProperty(
-     *      iri="http://pending.schema.org/headline"
-     * )
-     * @ORM\Column(type="string", length=512, nullable=false)
-     */
-    private $name;
-
-    /**
-     * @ApiProperty(
-     *      iri="http://pending.schema.org/about"
-     * )
-     * @ORM\OneToMany(targetEntity="App\Entity\Library\Book", mappedBy="serie")
-     * @ApiSubresource(maxDepth=1)
-     */
-    private $book;
-
-    /**
-     * Serie constructor.
-     */
-    public function __construct()
-    {
-        $this->book = new ArrayCollection();
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * @param mixed $name
-     * @return Serie
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getBook()
-    {
-        return $this->book;
-    }
-}
