+++ src/Entity/Review.php
@@ -1,0 +1,179 @@
+<?php
+namespace App\Entity;
+
+use ApiPlatform\Core\Annotation\ApiResource;
+use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Validator\Constraints as Assert;
+
+/**
+ * A review of a book.
+ *
+ * @ApiResource
+ * @ORM\Entity
+ */
+class Review
+{
+    /**
+     * @var int The id of this review.
+     *
+     * @ORM\Id
+     * @ORM\GeneratedValue
+     * @ORM\Column(type="integer")
+     */
+    private $id;
+
+    /**
+     * @var int The rating of this review (between 0 and 5).
+     *
+     * @ORM\Column(type="smallint")
+     * @Assert\Range(min=0, max=5)
+     */
+    private $rating;
+
+    /**
+     * @var string the body of the review.
+     *
+     * @ORM\Column(type="text")
+     * @Assert\NotBlank
+     */
+    private $body;
+
+    /**
+     * @var string The author of the review.
+     *
+     * @ORM\Column
+     * @Assert\NotBlank
+     */
+    private $author;
+
+    /**
+     * @var \DateTimeInterface The date of publication of this review.
+     *
+     * @ORM\Column(type="datetime")
+     * @Assert\NotBlank
+     */
+    private $publicationDate;
+
+    /**
+     * @var Book The book this review is about.
+     *
+     * @ORM\ManyToOne(targetEntity="Book", inversedBy="reviews")
+     * @Assert\NotNull
+     */
+    private $book;
+
+    /**
+     * @return int
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param int $id
+     * @return Review
+     */
+    public function setId($id)
+    {
+        $this->id = $id;
+
+        return $this;
+    }
+
+    /**
+     * @return int
+     */
+    public function getRating()
+    {
+        return $this->rating;
+    }
+
+    /**
+     * @param int $rating
+     * @return Review
+     */
+    public function setRating($rating)
+    {
+        $this->rating = $rating;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getBody()
+    {
+        return $this->body;
+    }
+
+    /**
+     * @param string $body
+     * @return Review
+     */
+    public function setBody($body)
+    {
+        $this->body = $body;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getAuthor()
+    {
+        return $this->author;
+    }
+
+    /**
+     * @param string $author
+     * @return Review
+     */
+    public function setAuthor($author)
+    {
+        $this->author = $author;
+
+        return $this;
+    }
+
+    /**
+     * @return \DateTimeInterface
+     */
+    public function getPublicationDate()
+    {
+        return $this->publicationDate;
+    }
+
+    /**
+     * @param \DateTimeInterface $publicationDate
+     * @return Review
+     */
+    public function setPublicationDate($publicationDate)
+    {
+        $this->publicationDate = $publicationDate;
+
+        return $this;
+    }
+
+    /**
+     * @return Book
+     */
+    public function getBook()
+    {
+        return $this->book;
+    }
+
+    /**
+     * @param Book $book
+     * @return Review
+     */
+    public function setBook($book)
+    {
+        $this->book = $book;
+
+        return $this;
+    }
+
+}
