+++ symfony.lock
@@ -1,4 +1,7 @@
 {
+    "api-platform/api-pack": {
+        "version": "v1.0.1"
+    },
     "api-platform/core": {
         "version": "2.1",
         "recipe": {
@@ -7,9 +10,6 @@
             "version": "2.1",
             "ref": "b50f8aa321cc40c380dda7605aa5cef66f476802"
         }
-    },
-    "clue/stream-filter": {
-        "version": "v1.4.0"
     },
     "doctrine/annotations": {
         "version": "1.0",
@@ -30,7 +30,7 @@
         "version": "v2.7.3"
     },
     "doctrine/dbal": {
-        "version": "v2.6.3"
+        "version": "v2.5.13"
     },
     "doctrine/doctrine-bundle": {
         "version": "1.6",
@@ -44,41 +44,17 @@
     "doctrine/doctrine-cache-bundle": {
         "version": "1.3.2"
     },
-    "doctrine/doctrine-migrations-bundle": {
-        "version": "1.2",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "master",
-            "version": "1.2",
-            "ref": "c1431086fec31f17fbcfe6d6d7e92059458facc1"
-        }
-    },
     "doctrine/inflector": {
         "version": "v1.2.0"
     },
     "doctrine/instantiator": {
-        "version": "1.1.0"
+        "version": "1.0.5"
     },
     "doctrine/lexer": {
         "version": "v1.0.1"
     },
-    "doctrine/migrations": {
-        "version": "v1.6.2"
-    },
     "doctrine/orm": {
-        "version": "v2.5.13"
-    },
-    "guzzlehttp/psr7": {
-        "version": "1.4.2"
-    },
-    "javiereguiluz/easyadmin-bundle": {
-        "version": "1.17",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "master",
-            "version": "1.17",
-            "ref": "66070ea553247b0cd5275e21d0fe5399a25f40be"
-        }
+        "version": "v2.5.12"
     },
     "jdorn/sql-formatter": {
         "version": "v1.2.17"
@@ -86,62 +62,11 @@
     "monolog/monolog": {
         "version": "1.23.0"
     },
-    "ocramius/package-versions": {
-        "version": "1.2.0"
-    },
-    "ocramius/proxy-manager": {
-        "version": "2.1.1"
-    },
-    "pagerfanta/pagerfanta": {
-        "version": "v1.0.5"
-    },
-    "php-http/cache-plugin": {
-        "version": "v1.5.0"
-    },
-    "php-http/client-common": {
-        "version": "1.7.0"
-    },
-    "php-http/curl-client": {
-        "version": "v1.7.0"
-    },
-    "php-http/discovery": {
-        "version": "1.3.0"
-    },
-    "php-http/httplug": {
-        "version": "v1.1.0"
-    },
-    "php-http/httplug-bundle": {
-        "version": "1.6",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "master",
-            "version": "1.6",
-            "ref": "4655962c64e2de96c418caf576c57215711c7877"
-        }
-    },
-    "php-http/httplug-pack": {
-        "version": "1.0.0"
-    },
-    "php-http/logger-plugin": {
-        "version": "v1.0.0"
-    },
-    "php-http/message": {
-        "version": "1.6.0"
-    },
-    "php-http/message-factory": {
-        "version": "v1.0.2"
-    },
-    "php-http/promise": {
-        "version": "v1.0.0"
-    },
-    "php-http/stopwatch-plugin": {
-        "version": "1.1.0"
-    },
     "phpdocumentor/reflection-common": {
         "version": "1.0.1"
     },
     "phpdocumentor/reflection-docblock": {
-        "version": "4.2.0"
+        "version": "4.1.1"
     },
     "phpdocumentor/type-resolver": {
         "version": "0.4.0"
@@ -152,17 +77,11 @@
     "psr/container": {
         "version": "1.0.0"
     },
-    "psr/http-message": {
-        "version": "1.0.1"
-    },
     "psr/log": {
         "version": "1.0.2"
     },
     "psr/simple-cache": {
         "version": "1.0.0"
-    },
-    "rebolon/api-pack": {
-        "version": "v1.0.1"
     },
     "sensio/framework-extra-bundle": {
         "version": "4.0",
@@ -174,7 +93,7 @@
         }
     },
     "symfony/asset": {
-        "version": "v4.0.1"
+        "version": "v3.3.13"
     },
     "symfony/cache": {
         "version": "v3.3.13"
@@ -198,7 +117,7 @@
         "version": "v3.3.13"
     },
     "symfony/doctrine-bridge": {
-        "version": "v4.0.1"
+        "version": "v3.3.13"
     },
     "symfony/dotenv": {
         "version": "v3.3.13"
@@ -207,7 +126,7 @@
         "version": "v3.3.13"
     },
     "symfony/expression-language": {
-        "version": "v4.0.1"
+        "version": "v3.3.13"
     },
     "symfony/filesystem": {
         "version": "v3.3.13"
@@ -223,9 +142,6 @@
             "version": "1.0",
             "ref": "e921bdbfe20cdefa3b82f379d1cd36df1bc8d115"
         }
-    },
-    "symfony/form": {
-        "version": "v4.0.1"
     },
     "symfony/framework-bundle": {
         "version": "3.3",
@@ -245,9 +161,6 @@
     "symfony/inflector": {
         "version": "v3.3.13"
     },
-    "symfony/intl": {
-        "version": "v4.0.1"
-    },
     "symfony/monolog-bridge": {
         "version": "v3.3.13"
     },
@@ -259,12 +172,6 @@
             "version": "3.1",
             "ref": "c24944bd87dacf0bb8fa218dc21e4a70fff56882"
         }
-    },
-    "symfony/options-resolver": {
-        "version": "v4.0.1"
-    },
-    "symfony/polyfill-intl-icu": {
-        "version": "v1.6.0"
     },
     "symfony/polyfill-mbstring": {
         "version": "v1.6.0"
@@ -291,7 +198,7 @@
         }
     },
     "symfony/security": {
-        "version": "v4.0.1"
+        "version": "v3.3.13"
     },
     "symfony/security-bundle": {
         "version": "3.3",
@@ -314,7 +221,7 @@
             "repo": "github.com/symfony/recipes",
             "branch": "master",
             "version": "3.3",
-            "ref": "6bcd6c570c017ea6ae5a7a6a027c929fd3542cd8"
+            "ref": "124be9a2b7cc035fd8c56f4fd2e19c907c1f3fb8"
         }
     },
     "symfony/twig-bridge": {
@@ -330,7 +237,7 @@
         }
     },
     "symfony/validator": {
-        "version": "v4.0.1"
+        "version": "v3.3.13"
     },
     "symfony/var-dumper": {
         "version": "v3.3.13"
@@ -356,15 +263,6 @@
     "symfony/yaml": {
         "version": "v3.3.13"
     },
-    "twig/extensions": {
-        "version": "1.0",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "master",
-            "version": "1.0",
-            "ref": "4851df0afc426b8f07204379d21fca25b6df5d68"
-        }
-    },
     "twig/twig": {
         "version": "v2.4.4"
     },
@@ -373,11 +271,5 @@
     },
     "willdurand/negotiation": {
         "version": "v2.3.1"
-    },
-    "zendframework/zend-code": {
-        "version": "3.3.0"
-    },
-    "zendframework/zend-eventmanager": {
-        "version": "3.2.0"
     }
 }
